Reformats code
