+++ app/src/main/java/com/wachon/spotiwrap/di/AppModule.kt
@@ -25,7 +25,7 @@ val FeaturesModule: Module
         includes(
             listOf(
                 LoginModule,
-                MenuModule, RecommenderModule,
+                MenuModule,
                 RecommenderModule,
                 SplashModule,
                 TracksModule,
